+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Allow empty strings when using the search command line option. For example `--search=""` or `-s=""`
 
 ## [v2.4.0] 2021-03-17
 ### Changed